+++ src/lmos/__init__.py
@@ -1,3 +1,5 @@
-
 from . import _version
-__version__ = _version.get_versions()['version']
+from .lmos import (median_of_squares, minimize_multiple_guesses,
+                   numerically_optimize)
+
+__version__ = _version.get_versions()["version"]